voice_call_neymar/requirements.txt
# LiveKit Agents with required plugins
livekit-agents>=0.12.0
livekit-plugins-groq>=0.3.0
livekit-plugins-elevenlabs>=0.8.0
livekit-plugins-silero>=0.7.0

# Environment and utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
